Change data Key type from size_t to int to fix a test failure using gcc on Linux 64bits
git-svn-id: https://stlport.svn.sourceforge.net/svnroot/stlport/branches/STLPORT_5_1/STLport@2530 01b27279-081b-0410-8cf9-d9d9e0fb0389

+++ test/unit/unordered_test.cpp
@@ -464,16 +464,16 @@ struct Key
   Key() : m_data(0) {}
   explicit Key(size_t data) : m_data(data) {}
 
-  size_t m_data;
+  int m_data;
 };
 
 struct KeyHash
 {
   size_t operator () (Key key) const
-  { return key.m_data; }
+  { return (size_t)key.m_data; }
 
-  size_t operator () (size_t data) const
-  { return data; }
+  size_t operator () (int data) const
+  { return (size_t)data; }
 };
 
 struct KeyEqual
@@ -481,20 +481,20 @@ struct KeyEqual
   bool operator () (Key lhs, Key rhs) const
   { return lhs.m_data == rhs.m_data; }
 
-  bool operator () (Key lhs, size_t rhs) const
+  bool operator () (Key lhs, int rhs) const
   { return lhs.m_data == rhs; }
 
-  bool operator () (size_t lhs, Key rhs) const
+  bool operator () (int lhs, Key rhs) const
   { return lhs == rhs.m_data; }
 };
 
 struct KeyHashPtr
 {
   size_t operator () (Key const volatile *key) const
-  { return key->m_data; }
+  { return (size_t)key->m_data; }
 
-  size_t operator () (size_t data) const
-  { return data; }
+  size_t operator () (int data) const
+  { return (size_t)data; }
 };
 
 struct KeyEqualPtr
@@ -502,10 +502,10 @@ struct KeyEqualPtr
   bool operator () (Key const volatile *lhs, Key const volatile *rhs) const
   { return lhs->m_data == rhs->m_data; }
 
-  bool operator () (Key const volatile *lhs, size_t rhs) const
+  bool operator () (Key const volatile *lhs, int rhs) const
   { return lhs->m_data == rhs; }
 
-  bool operator () (size_t lhs, Key const volatile *rhs) const
+  bool operator () (int lhs, Key const volatile *rhs) const
   { return lhs == rhs->m_data; }
 };
